requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiobotocore==2.21.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
awscli==1.38.9
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.9.0
boto3==1.37.9
botocore==1.37.9
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
Django==5.1.6
djangorestframework==3.15.2
dm-tree==0.1.9
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.16
etils==1.12.0
executing==1.2.0
fastapi==0.115.11
fastjsonschema==2.16.2
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
geos==0.2.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.29.2
hydra-core==1.3.2
idna==3.4
immutabledict==4.2.1
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipykernel==6.21.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.0.2
jupyter-console==6.5.0
jupyter_core==5.2.0
jupyter-events==0.6.3
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lightning-utilities==0.13.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.4.1
mock==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
namex==0.0.8
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.4.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.14.1
packaging==23.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.3
peft==0.14.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pip==23.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.36
propcache==0.3.0
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.14.0
pyparsing==3.2.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.5
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2024.11.6
requests==2.32.3
reverse-geocoder==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rsa==4.7.2
s3fs==0.4.2
s3transfer==0.11.4
safetensors==0.5.3
sagemaker==2.241.0
sagemaker-core==1.0.25
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==65.5.0
simple-parsing==0.1.7
six==1.16.0
smart-open==7.1.0
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.5.3
stack-data==0.6.2
starlette==0.46.1
sympy==1.13.1
tblib==3.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchmetrics==1.6.2
torchvision==0.21.0
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.2.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
wandb==0.19.8
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.5
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
